+++ src/simulation/elements/PRTI.cpp
@@ -58,7 +58,7 @@ int Element_PRTI::update(UPDATE_FUNC_ARGS)
 {
 	int fe = 0;
 
-	parts[i].tmp = (int)((parts[i].temp-73.15f)/100+1);
+	parts[i].tmp = (int)(((float)parts[i].temp-73.15f)/100+1);
 	if (parts[i].tmp >= CHANNELS)
 		parts[i].tmp = CHANNELS-1;
 	else if (parts[i].tmp < 0)
