+++ src/simulation/elements/QRTZ.cpp
@@ -53,7 +53,7 @@ int Element_QRTZ::update(UPDATE_FUNC_ARGS)
 	{
 		parts[i].pavg[0] = parts[i].pavg[1];
 		parts[i].pavg[1] = sim->pv[y/CELL][x/CELL];
-		if (parts[i].pavg[1]-parts[i].pavg[0] > 0.05*(parts[i].temp/3) || parts[i].pavg[1]-parts[i].pavg[0] < -0.05*(parts[i].temp/3))
+		if (parts[i].pavg[1]-parts[i].pavg[0] > 0.05*((float)parts[i].temp/3) || parts[i].pavg[1]-parts[i].pavg[0] < -0.05*((float)parts[i].temp/3))
 		{
 			sim->part_change_type(i,x,y,PT_PQRT);
 			parts[i].life = 5; //timer before it can grow or diffuse again
